+++ CHANGES.md
@@ -8,9 +8,9 @@
 * Support 8-bit colour palette, bright colours. (#6, @MisterDA)
 
 * Expose default foreground and background colors with `.fg-default`
-  and `.bg-default` CSS classes. (#6, MisterDA)
+  and `.bg-default` CSS classes. (#6, #10, MisterDA)
 
-* Fix bold/bright confusion, expose `.bold` CSS class. (#6, @MisterDA)
+* Fix bold/bright confusion, expose `.bold` CSS class. (#6, #10 @MisterDA)
 
 * Expose alternative 16-colours colorschemes (xterm dark, solarized
   light and dark). (#6, @MisterDA)