+++ CHANGES.md
@@ -1,4 +1,4 @@
-### unreleased
+### 0.6.0
 
 * Support OSC hyperlinks. (#13, @MisterDA)
   https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda